ai_requirements.txt
tensorflow==2.15.0
tensorflow-cpu==2.15.0
numpy==1.24.3
Pillow==10.0.1
tqdm==4.66.1
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
scikit-learn==1.3.0
opencv-python==4.8.1.78
h5py==3.10.0
psutil==5.9.5
memory-profiler==0.61.0 